requirements.txt
discord.py==2.3.2
pathlib==1.0.1
python-dotenv==1.0.0
Wavelink==3.1.0
nest-asyncio==1.5.6